+++ test/runtests.jl
@@ -14,8 +14,13 @@ cd("..")
     @test se().relaxation == 0.4
     @test se().elevation_rate == 0.0
     @test se().azimuth_rate == 0.0
-    @test se().l_tether == 50.0
-    @test se().v_reel_out == 0.0
+    set = deepcopy(se())
+    @test set.l_tether == 50.0
+    set.l_tether = 51.0
+    @test set.l_tether == 51
+    @test set.v_reel_out == 0.0
+    set.v_reel_out = 1.0
+    @test set.v_reel_out == 1.0
     @test se("system2.yaml").cs_4p == 1.1
     @test length(se().alpha_cl) == 12
     set_data_path(tempdir())